Initial Page
This is the initial page for the drumkit. The styling is a work-in-progress, but
I want it to have a very retro look. Perhaps that'll be the theme of this 30 day
project. Retro, hot but with the old mac os dialog boxes. Anyway, we'll
implement the drumkit stuff using vanilla javascript and we'll have some audio
files for the drum sounds.

drumkit.html:
        - The html page
        - I'll probably change as we add more classes and attributes in order to
          implement the drumkit

style.css:
        - The styling for this page