Adding Javascript and New Styles
I'm sure of the aesthetic of the styling but I'm still unsure how to I want to
manipulate the space. In addition, I have an idea of how I want the layout for
smaller to look, however I still want to experiment some more before committing
to it.

The javascript took a little work. I just wasn't sure how to do it exactly. At
first, I thought it would be easy; I thought could listen for keydown events on
each box. However, that didn't make any sense because we would have to somehow
focus on a specific box to register a key press with it so that the browser
would fire an event and cause the box's handler to be invoked. We can't focus on
an individual element like we can links, so this method won't work.

Instead, we'll data attributes to the HTML and assign them the keycode, an
integer value representing a key, corresponds to the key described in the page.
In the script, we'll take the keydown event's keycode -- the listener for
keydown events is registered with the document -- and use it to query the page
for an element whose data attribute's value matches the keycode. We do this by
creating a template string, representing a selector, and interpolating the
event's keycode into the string when invoked -- the function is invoked when the
event is captured. This is one way we can associate a keydown event with an
element in the page.

Once an element has been matched to the selector, we can add a class that
corresponds to a style of a pressed key. This will cause the browser to
apply the style immediately once the key is pressed.

drumkit.html:
        - Replacing p's with kbd; it's better description of what these
          boxes actually are: keys.
        - Replacing p's with span; again, this more accurately describes the
          box's label as it isn't a paragraph.
        - data- attributes are added to the key elements. These attributes are
          assigned the keycode values so that we can associate an element with
          a keydown event via the keycode.
drumkit.js:
        - Adding a keydown listener to the document
        - Keydown events are matched with the correct element using data
          attributes.
drumkit.css:
        - Drop shadows for when a key is pressed
                - Later on, we'll remove class so that it looks like the style
                  is only applied when the key is actually pressed.
        - Shortened the boxes to resemble keys
        - The layout is still in progress, e.g, the spacing, styling for smaller
          viewports, etc.