Removing Drop Shadows and Prelude to Audio
We would like to simulate the keydown-keyup behavior on screen but adding and
removing a drop shadow on the box (or removing and adding a drop shadow
depending on your tastes or user experience). However, once we add a
class to an element, which we currently are doing, the style described in the
class definition is applied and persists because the class is now attached to
the element. What we want is to remove the drop shadow after the key comes up.
We could add a listener to remove the class once the key comes up, but since
we're applying a transition effect, that transition might be abruptly
interrupted if the event is fired before transition completes, or there is a
delay in removing the style if the event's handler is queued up and isn't
invoked in a timely manner -- we're mostly concerned about the former situation.

Instead of relying on the key up event, we'll rely on the fact that transitions
cause the element to fire events once its transition completes. This way we
allow the transition to complete before removing the class that contains the
unwanted styling.

Removing the styling is slightly different that what we did to add it. When we
added the styling, we searched the DOM for the div that corresponded to the key
pressed and added the class to the attributes. For removing the class, we
immediately add listeners to all of the divs so that they can start listening for
'transitionend' before any keys are pressed -- we don't start listening key
press events until after we established this code. This way we don't miss any
fired events.

I found something interesting about querySelector and getElementsbyClass.
Initially, I thought we could use getElementbyClass and use a forEach function
to iterate through the list. However, I found that getElementByClass doesn't
support the forEach function because the object returned, HTMLCollection,
doesn't have a forEach method. Instead, I had to use querySelectorAll which
returns a non-live NodeList (a list of nodes) and has a forEach implementation.
Sure, I could have iterator through the HTMLCollection and added listeners that
way, but it looks cleaner using forEach. Note that I don't need a live list
of elements so using querySelectorAll is alright.

NOTE: Note all browsers implement the forEach method for NodeList. In this case,
you'll have to iterate using a for loop and indices or use Array.from method to
convert the list into an array.

As previously alluded to, HTMLCollection returns a collection that acts like a
list of live DOM elements. (It seems HTMLCollection is a special case of
NodeList? Though it seems not since HTMLCollection doesn't inherit NodeList
stuff; if it was a special case, it would. No, I think NodeList is special case
of HTMLCollection as it implements its own methods and changes the element in
the list from live to non-live). Unlike HTMLCollection, querySelectorAll doesn't
return a list of live DOM elements, rather it takes snapshots of them. Yet, both
contain Nodes.

With Nodes, you can add event listeners and other stuff.

Finally, another interesting fact is that a transition on an element can be
composed of multiple transitions. We might change all sides of the border, the
background, etc. All of these individual transitions emit a transitionend event.
Some of this is hinted by the fact the CSS rule calls for us to explicitly name
a property that is will undergo a transition along with the desired settings, or
we might apply the same settings for all properties undergoing a change.

drumkit.html:
        - Adding the sounds to the html source
                - They're semantically described using HTML5's audio tags.
                - They're also given the data-key attribute, an HTML standard
                  used to store data associated with a particular element, but
                  has no defined meaning, to associate a key press with this
                  audio element via the key code. This is the exact method we
                  used to associate key events with the divs of class key.
drumkit.js:
        - We now remove the class that was added during a keydown event. This
          will simulate a key press on screen so as to give feedback to the
          user.
style.css:
        - Changed the background of the box when the key is pressed.
/sounds:
        - The drum sounds that'll be played when a user hits a key.