Adding and Play the Drum Audio
Now we put our audio to use. We'll use the same method that we used to associate
a div on the page to a key press, but, now, we'll associate an audio element
with a key event.

We use the data-key attributes to store the keycode on the element. In the
script, when a key event occurs, we use the keycode in the event to search for
the audio element that matches the data-key value. To search for the element, we
use the querySelector() method and pass in a template string representing the
selector; we interpolate the keycode into the selector string once the key event
occurs. This will allow us to inject the correct code into the selector so that
we can search for the element that has an attribute and attribute value matches
it.

Once we grab the audio DOM element, we can invoke the play method, which audio
DOM elements implement. This will play the sound defined in the element.

drumkit/drumkit.js:
        - Plays the drumkit sound associated with the key pressed.