Syncing the Transition Duration and Rewinding Sound
The transition duration is too long which discourages quick, repeated key
presses. The long duration feels as if we have to wait for the audio to complete
before we can make another key press. Sometimes the box styles get stuck with
repeated presses because of the long transition durations. Therefore,
adjustments are made to the transition duration so that the transition animation
completes quickly before another transition occurs. This gives the feeling that
repeated key presses have registered rather than there being a delay because the
transition hasn't completed.

On the script side, the behavior of repeated calls to play() is to wait for a
previous play to finish before playing the file again. This poses a problem
because we want key presses to interrupt the current sound and play a new
instance of it. Therefore, in order to accomplish this, we have to essentially
rewind or, in digital terms, move the marker to the beginning and play the sound
over again. We do this by setting the currentTime to 0 when a key event occurs.
(It does seem like the sounds play simultaneously if we hit multiple keys at
once. Different sounds don't interrupt each other which means that the
currentTime marker is moved only for that specific audio file. This makes
perfect sense if you think about it: the currentTime is a property of that
specific audio DOM element.)

drumkit/drumkit.js:
        - We move the playing marker for a particular sound back to beginning
          for each key press. This will interrupt the playback of the
          corresponding audio file and play it from the beginning. Successive
          key presses create sound immediately rather than waiting for each
          key event to finish playing its sound before it can be played again.
drumkit/style.css:
        - Adjusting the transition so that the effect syncs better with the
          sound.