Moving the Styling to Mobile First
I couldn't get a grid of boxes for viewports between 400px and 600px by using
flex-wrap.  Instead of creating multiple media queries for each time the grid
broke down, we make the mobile-first layout the base layout. This style would
persist until too much space was alloted on the left and right hand sides of the
grid. It is at this point when we add rules to relax the layout a bit since
there is more space for keys to fit; a grid isn't necessary.

When the viewport finally reaches 1000px, the layout is relaxes so that all the
keys can line up in a row.

Flex-basis is used on the container -- which is also a flex-item -- to force the
keys, using flex wrap, into a 3x3 grid. As the viewport increases, the
flex-basis is expanded enough to let more keys fit in the first row until,
finally, all keys fit in a single row.

drumkit/style.css:
        - Switching to a mobile-first layout
        - As the viewport expands, The layout allows for more keys to fit on the
          top row until all keys fit on a single line.