Clock Project
This is a simple clock styled after those retro wall clocks back in the 1950s.
Groovy huh?! Currently, it doesn't look like much. We have all the main pieces
setup.

Javascript is going to be used to apply the transformations on the hand elements
based on the current time. The idea is to use the Date constructor to get the
current date and, using some math, map the time to the appropriate degree in the
circle (clock face). We're going to have to use to some math to do the mapping
and we'll encounter some challenges in creating this mapping due to how the
rotate functions in CSS are implemented.

I'm not exactly sure how we'll continuously update the angle (or apply the CSS
rotate function for that matter) just yet. Also, how would we continuously get
the current time? To be determined...

clock.html:
        - Basic page consisting of a few divs
clock.js:
        - Nothing just yet
style.css:
        - The styling for the page. I didn't run into too much trouble here.
          It's a pretty straightforward styling. It uses flexbox though.