Moving the Minute Hand Per Second Within a 60 Minute Period
Instead of having the minute hand move 6 degrees for every minute elapsed, we'll
"inch" forward the hand for every second in a 60 minute period. This will give
off a mechanical animation as if it's in sync with the seconds hand: the minute
hand will move 360 deg in 3600seconds (3600seconds in an hour). This gives us a
rate of 1/10th of a degree per second.  To get the correct position, we move the
minute hand 1/10th of a degree for each of the elapsed number of second (60secs
* current_min + current_secs).

We need the current amount of seconds elapsed because this function is invoked
every second for every minute; the expression 60seconds*current_minute will
always give us the same result within a 60 second period because the minute
never changes. The expression converts the number of minutes into seconds.
Therefore, get the current total number of seconds elapsed, we'll have add the
number of second elapsed within the span of the current minute.

The expression below results in the number of degrees that the hand should move
given the current amount of seconds elapsed within a span of an hour.

I'll be sure to add the same behavior to the hour hand.

I also need to figure out how to optimize this. Calculating and drawing this
stuff this frequently has to be taxing for something that just tells time.

clock/clock.js:
        - Changing the minute hand's behavior. Previously the hand would move
          to reflect the current minutes. Now we "inch" forward the hand every
          second. This gives the clock a mechanical feeling as if the minute
          hand is in sync with the seconds hand.