Refactoring a Few Things
Querying for the clock hands every second is excessive; it should be enough to
query the document for them once it has loaded, which is why I'm moving the
statements out of the draw function.

Oh, and I've renamed the moveSeconds function to draw.

I was also creating new Date objects for each of the time properties that I
wanted to extract. I'm fixing that as of this commit. One date object is being
created at the time of draw (one date object every second).  The time values we
need aren't changing as we draw, therefore we don't need the precise time at the
execution of the statement.

I've also removed some comments which were cluttering up the function.

clock/clock.js:
        - The document is queried only once the DOM has loaded instead of every
          second.
        - Only one Date object is created instead of 3 of them.
        - Renaming some of the inner functions.
        <meta charset='utf-8'>